python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.7",
    "langchain-groq>=0.2.3",
    "langchain-ollama>=0.2.2",
    "parsera>=0.2.2",
    "patchright>=1.49.1",
    "pytest-playwright>=0.6.2",
    "setuptools>=75.8.0",
    "uvicorn>=0.34.0",
]
